# python/services/openai_plugin/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "All in One Plugin for developers",
        "description": "A plugin that connects tools for developers into ChatGPT",
        "version": "v1"
    },
    "servers": [
        {
            "url": "https://aio.fly.dev/openai-plugin"
        }
    ],
    "paths": {
        "/github/repository/{owner}/{repo}": {
            "get": {
                "operationId": "getGithubRepository",
                "summary": "Get a full repository from GitHub with its files and contents.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "owner",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "The owner of the repository."
                    },
                    {
                        "in": "path",
                        "name": "repo",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "The name of the repository."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/getGithubRepository"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "getGithubRepository": {
                "type": "array",
                "items": {
                    "type": "object",
                    "properties": {
                        "path": {
                            "type": "string"
                        },
                        "content": {
                            "type": "string"
                        }
                    }
                },
                "description": "The list of files in the repository."
            }
        }
    }
}